refactor: graphqlディレクトリを削除
