--- renderer/src/graphql/queries/index.ts
@@ -1,2 +1,0 @@
-export * from './LoginUserQuery';
-export * from './SearchPullRequestsQuery';
